+++ event-manager/src/main/java/ru/haritonenko/eventmanager/event/api/dto/EventUpdateRequestDto.java
@@ -23,6 +23,6 @@ public record EventUpdateRequestDto(
         Integer duration,
         @NotNull(message = "Event location id can not be null")
         @Min(value = 1, message = "Min location id is 1")
-        Integer locationId
+        Long locationId
 ) {
 }
