+++ event-manager/src/main/java/ru/haritonenko/eventmanager/kafka/producer/configuration/KafkaProducerConfiguration.java
@@ -1,7 +1,7 @@
 package ru.haritonenko.eventmanager.kafka.producer.configuration;
 
 import org.apache.kafka.clients.producer.ProducerConfig;
-import org.apache.kafka.common.serialization.IntegerSerializer;
+import org.apache.kafka.common.serialization.LongSerializer;
 import org.springframework.boot.autoconfigure.kafka.KafkaProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -16,15 +16,15 @@ import java.util.Map;
 public class KafkaProducerConfiguration {
 
     @Bean
-    public DefaultKafkaProducerFactory<Integer, EventChangeKafkaMessage> producerFactory(KafkaProperties kafkaProperties) {
+    public DefaultKafkaProducerFactory<Long, EventChangeKafkaMessage> producerFactory(KafkaProperties kafkaProperties) {
         Map<String, Object> props = kafkaProperties.buildProducerProperties();
-        props.put(ProducerConfig.KEY_SERIALIZER_CLASS_CONFIG, IntegerSerializer.class);
+        props.put(ProducerConfig.KEY_SERIALIZER_CLASS_CONFIG, LongSerializer.class);
         props.put(ProducerConfig.VALUE_SERIALIZER_CLASS_CONFIG, JsonSerializer.class);
         return new DefaultKafkaProducerFactory<>(props);
     }
 
     @Bean
-    public KafkaTemplate<Integer, EventChangeKafkaMessage> kafkaTemplate(DefaultKafkaProducerFactory<Integer, EventChangeKafkaMessage> producerFactory) {
+    public KafkaTemplate<Long, EventChangeKafkaMessage> kafkaTemplate(DefaultKafkaProducerFactory<Long, EventChangeKafkaMessage> producerFactory) {
         return new KafkaTemplate<>(producerFactory);
     }
 }
